Version 0.0.1.0: Water sound effects added, Functional & Unit Test exploring begins
Lots of stuff added here, with the most noticeable changes being that I added a number of sound effects that are triggered to sound when the water hose is sprayed. I began playing around with setting up functional tests too. I also added a 'showfps' feature to the character blueprint, and delved into some performance and profiling features.